Move FirebaseRefs and ListID to new util package
