+++ app/src/main/java/kres/realtimeshoppinglist/firebase/productList/ProductListManager.java
@@ -2,7 +2,7 @@ package kres.realtimeshoppinglist.firebase.productList;
 
 import com.google.firebase.database.DatabaseReference;
 
-import kres.realtimeshoppinglist.firebase.FirebaseRefs;
+import kres.realtimeshoppinglist.firebase.util.FirebaseRefs;
 import kres.realtimeshoppinglist.model.Product;
 
 public class ProductListManager {
