+++ app/src/main/java/kres/realtimeshoppinglist/firebase/util/FirebaseRefs.java
@@ -1,4 +1,4 @@
-package kres.realtimeshoppinglist.firebase;
+package kres.realtimeshoppinglist.firebase.util;
 
 import android.provider.ContactsContract;
